Introduce --pulse-fix-mode.
This replaces the previous --pulse-function-only option.

Usage: The three options --pulse-fix-mode --pulse-fix-file=<file>
and --pulse-fix-function=<func> should be used together. This mode
should only be invoked after a whole-analysis was already done on the
entire project.

This fix mode only analyze the ONE specified function, and not its
dependencies. The unfiltered summary (in terms of path condition
and state on each disjunct) is written out as json.
The reporting phase is also disabled in this mode to save some time.